reorganized some files (#1210)
