+++ gradio/__init__.py
@@ -4,7 +4,8 @@ import gradio.components as components
 import gradio.inputs as inputs
 import gradio.outputs as outputs
 import gradio.processing_utils
-from gradio.blocks import Blocks, Column, Row, TabItem, Tabs
+from gradio.blocks import Blocks
+from gradio.layouts import Column, Row, TabItem, Tabs
 from gradio.components import (
     HTML,
     JSON,
